% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggNodeAttr.R
\name{aggNodeAttr}
\alias{aggNodeAttr}
\title{Node Attribute Aggregater}
\usage{
aggNodeAttr(node_df, measure_of_cent = "mean")
}
\arguments{
\item{node_df}{A data frame containing all the characteristics of the nodes
in the network. If there are n networks, a maximum of x nodes per network
and y variables for each node, the data frame should have n rows and x*y
columns. The column names of each variable should be written as follows: 
var1, var2, ... , varX.}

\item{measure_of_cent}{A vector that contains the measures of centrality with
which to summarize the node attributes. The supported measures are "mean"
and "median". Note that missing values are excluded from the calculations.}
}
\description{
\code{aggNodeAttr} creates a data frame that summarizes node attributes.
}
\author{
Philippe Boileau , \email{philippe_boileau@berkeley.edu}
}
