% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcICLNoPlots.R
\name{calcICLNoPlots}
\alias{calcICLNoPlots}
\title{Consensus Cluster Plus without Plots}
\usage{
calcICLNoPlots(consensus_results)
}
\arguments{
\item{consensus_results}{Results of consensus clustering. The second item
in the list returned by \code{\link{neatmap}}.}
}
\description{
\code{consensusClusterNoPlots} is a wrapper function for
  \code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}} that suppresses
  the creation of the plots that are created automatically.
}
\author{
Philippe Boileau , \email{philippe_boileau@berkeley.edu}
}
