% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterList.R
\docType{data}
\name{clustList}
\alias{clustList}
\title{Example of Cluster List}
\format{An object of class \code{list} of length 2.}
\usage{
clustList
}
\description{
This list represents an example of what the neatmap
function outputs as a list of significant clusters.
}
\author{
Phil Boileau, \email{philippe.boileau@mail.concordia.ca}
}
\keyword{datasets}
