% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactAlterAttr.R
\name{compactAlterAttr}
\alias{compactAlterAttr}
\title{Alter Attribute Compactor}
\usage{
compactAlterAttr(alter.df, measureOfCent = "mean")
}
\arguments{
\item{alter.df}{A data frame containing all the variables that are linked to
the alters in the network. If there are n networks, a maximum of x alters
per network and y variables for each alter, the data frame should have n
rows and x*y columns. The column names a shared attribute should be writen
as follows: attr1, attr2, ... , attrx.}

\item{measureOfCent}{A vector that contains the measures of centrality with
which to summarize the alter attributes. The supported measures are "mean"
and "median". Defaults to "mean".}
}
\description{
\code{compactAlterAttr} creates a data frame that summarizes alter attributes.
}
