% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactNodeAttr.R
\name{compactNodeAttr}
\alias{compactNodeAttr}
\title{Node Attribute Compactor}
\usage{
compactNodeAttr(node.df, measureOfCent = "mean")
}
\arguments{
\item{node.df}{A data frame containing all the characteristics of
the nodes in the network. If there are n networks, a maximum of x nodes
per network and y variables for each node, the data frame should have n
rows and x*y columns. The column names of each variable should be writen
as follows: var1, var2, ... , varX.}

\item{measureOfCent}{A vector that contains the measures of centrality with
which to summarize the node attributes. The supported measures are "mean"
and "median".}
}
\description{
\code{compactNodeAttr} creates a data frame that summarizes node attributes.
}
\author{
Phil Boileau , \email{philippe.boileau@mail.concordia.ca}
}
