% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_df.R
\docType{data}
\name{edge_df}
\alias{edge_df}
\title{Edge List Data Frame}
\format{An object of class \code{data.frame} with 10 rows and 10 columns.}
\usage{
edge_df
}
\description{
A dataset containing a list of undirected edges for ten different networks.
 Each network has a maximum size of 5 nodes. The network and node attribute
 data are saved in their respective files.
}
\author{
Philippe Boileau, \email{philippe_boileau@berkeley.edu}
}
\keyword{datasets}
