% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeData.R
\docType{data}
\name{edgeDF}
\alias{edgeDF}
\title{Edge List Data}
\format{An object of class \code{data.frame} with 10 rows and 10 columns.}
\usage{
edgeDF
}
\description{
A dataset containg a list of undirected edges
for ten different networks. Each network has 
a maximum size of 5 nodes. The network and 
node attribute data are saved in their
resepective files.
}
\author{
Phil Boileau, \email{philippe.boileau@mail.concordia.ca}
}
\keyword{datasets}
