% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egonetDataFrame.R
\name{egonet.data.frame}
\alias{egonet.data.frame}
\title{Ego Network Data Frames}
\usage{
egonet.data.frame(net.attr.df, alter.attr.df, edge.df,
  measuresOfCent = c("mean"))
}
\arguments{
\item{net.attr.df}{A data frame consisting of all of the ego networks' graph
attributes. The first column should contain the name of the network, and all
other columns should be numeric. All empty entries should be filled as "NA".}

\item{alter.attr.df}{A data frame consisting of all of the ego networks' alters'
attributes. All columns should be numeric. All empty entries should be filled
in as "NA".}

\item{edge.df}{A dataframe consisting of the edge matrix for each ego network.
Edges are assumed to be undirected and unweighted. 1 indicates the existence 
of an edge between alters, 0 indicates the lack of an edge.}

\item{measuresOfCent}{A vector of the measures of centrality to be used for the
summary of the alter attributes data. The supported measures of centrality
are: "mean", "mode" and "median". Defaults to "mean".}
}
\value{
The function returns a data frame that offers an overview of all of the
  ego networks.
}
\description{
\code{egonet.data.frame} produces data frames of collections of ego networks.
}
\details{
The function produces data frames of collections of ego networks. The function
requires the input of three data frames: a data frame containing the graph
attributes, a data frame containing the alter characteristics and a data frame
containing the edgelist of each network. The rows in each of these data frames
must represent individual ego networks, and must therefore have identical row
length. Measures of centrality used in the summarization of the alter attributes
must also be furnished. If users wish to analyze the homophily of alters in the 
ego networks, the variables to be analyzed must be indicated along with the
measures of homophily to be used. All of the ego networks should be of similar
size.
}
