% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatClusters.R
\name{formatCluster}
\alias{formatCluster}
\title{Format Cluster Output}
\usage{
formatCluster(clusterList)
}
\arguments{
\item{clusterList}{The list containing all the clusters of the 
hierarchical clustering.}
}
\description{
\code{formatCluster} formats the output of the cluster analysis.
}
\examples{
# dummy cluster results
clustList <- list(c("A", "B"), c("C", "D", "E"))
formatCluster(clusterList = clustList)
}
\author{
Phil Boileau , \email{philippe.boileau@mail.concordia.ca}
}
