% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{hierarchy}
\alias{hierarchy}
\title{Hierarchy}
\usage{
hierarchy(egonet)
}
\arguments{
\item{egonet}{An igraph object representing an ego network}
}
\description{
\code{hierarchy} calculates the hierarchy of an ego network
}
