% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatmaps.R
\docType{package}
\name{neatmaps}
\alias{neatmaps}
\alias{neatmaps-package}
\title{\code{neatmaps} package}
\description{
A package for exploring multi-network data.
}
\details{
See the README on
\href{ https://CRAN.R-project.org/package=neatmaps}{CRAN}
or \href{https://github.com/PhilBoileau/neatmaps/tree/master}{GitHub}
}
\author{
Philippe Boileau, \email{philippe_boileau@berkeley.edu}
}
