% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_attr_df.R
\docType{data}
\name{network_attr_df}
\alias{network_attr_df}
\title{Network Attributes Data}
\format{An object of class \code{data.frame} with 10 rows and 4 columns.}
\usage{
network_attr_df
}
\description{
A data set containing four randomly generated variables used to mimick the 
network attributes of ten different networks. Attribute 1 and 2 have a 
correlation of 0.41, 1 and 3 have a correlation of 0.91, 1 and 4 have a
correlation of 0.34, 2 and 3 have a correlation of 0.07, 2 and 4 have a
correlation of 0.17 and 3 and 4 have a correlation of 0.32.
}
\author{
Philippe Boileau, \email{philippe_boileau@berkeley.edu}
}
\keyword{datasets}
