% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkAttr.R
\docType{data}
\name{networkAttrDF}
\alias{networkAttrDF}
\title{Network Attributes Data}
\format{An object of class \code{data.frame} with 10 rows and 4 columns.}
\usage{
networkAttrDF
}
\description{
A data set containing four randomly generated variables
used to mimick the network attributes of ten different 
networks. Attribute 1 and 2 have a correlation of 0.41,
1 and 3 have a correlation of 0.91, 1 and 4 have a
correlation of 0.34, 2 and 3 have a correlation of 0.07,
2 and 4 have a correlation of 0.17 and 3 and 4 have a
correlation of 0.32.
}
\author{
Phil Boileau, \email{philippe.boileau@mail.concordia.ca}
}
\keyword{datasets}
