% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_attr_df.R
\docType{data}
\name{node_attr_df}
\alias{node_attr_df}
\title{Node Attribute Data}
\format{An object of class \code{data.frame} with 10 rows and 25 columns.}
\usage{
node_attr_df
}
\description{
A dataset containing randomly generated node attributes for each node in each
of the ten networks. Attributes A and B have a correlation of roughly 0.8, A
and C have a correlation of roughly -0.2, B and C have a correlation of
roughly 0.5. Attributes D and E were generated completely randomly, and
should not be strongly correlated with any of the other attributes.
}
\author{
Philippe Boileau, \email{philippe_boileau@berkeley.edu}
}
\keyword{datasets}
