% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeAttr.R
\docType{data}
\name{nodeAttrDF}
\alias{nodeAttrDF}
\title{Node Attribute Data}
\format{An object of class \code{data.frame} with 10 rows and 25 columns.}
\usage{
nodeAttrDF
}
\description{
A dataset containing randomly generated node attributes
for each node in each of the ten networks. Attributes A
and B have a correlation of roughly 0.8, A and C have a
correlation of roughly -0.2, B and C have a correlation
of roughly 0.5. Attributes D and E were generated
completely randomly, and should not be strongly correlated
with any of the other attributes.
}
\author{
Phil Boileau, \email{philippe.boileau@mail.concordia.ca}
}
\keyword{datasets}
