% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledColumns.R
\name{scaleColumns}
\alias{scaleColumns}
\title{Scale Between 0 and 1}
\usage{
scaleColumns(df)
}
\arguments{
\item{df}{The data frame of numerical values to be scaled.}
}
\description{
\code{scaleColumns} scales the columns of a data frame object
  between the values of 0 and 1 without changing the underlying
  distribution of the columns.
}
