% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructuralAttr.R
\name{getStructureAttr}
\alias{getStructureAttr}
\title{Structural Attributes of Networks Data Frame}
\usage{
getStructureAttr(net_list)
}
\arguments{
\item{net_list}{A list of Igraph network objects that represent the collection
of networks.}
}
\description{
\code{getStructureAttr} produces a data frame of the structural attributes
  of a list of networks.
}
\author{
Philippe Boileau, \email{philippe_boileau@berkeley.edu}
}
