% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructuralAttr.R
\name{getStructureAttr}
\alias{getStructureAttr}
\title{Structural Attributes of Egonets Data Frame}
\usage{
getStructureAttr(egonetList)
}
\arguments{
\item{egonetList}{A list of Igraph network objects that represent the collection
of egonets.}
}
\description{
\code{getStructureAttr} produces a data frame of the structural attributes
  of a list of ego networks.
}
