% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{hierarchy}
\alias{hierarchy}
\title{Hierarchy}
\usage{
hierarchy(net)
}
\arguments{
\item{net}{An igraph object representing a network}
}
\description{
\code{hierarchy} calculates the hierarchy of a network
}
\author{
Phil Boileau , \email{philippe.boileau@mail.concordia.ca}
}
